+++ Qlik.Sense.RestClient/ExamplesQcs/ListAllApps/Program.cs
@@ -1,8 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Newtonsoft.Json.Linq;
 using Qlik.Sense.RestClient;
 
 namespace ListAllApps
@@ -13,32 +9,20 @@ namespace ListAllApps
         {
             var url = "<url>";
             var apiKey = "<apiKey>";
-
+            
             var restClient = new RestClientQcs(url);
             restClient.AsApiKey(apiKey);
 
+            Console.WriteLine("Fetching all apps:");
+            var appCnt = 0;
             // Fetch app items using page size 20 (maximum is 100).
-            var allApps = FetchAll(restClient, "/api/v1/items?resourceType=app&limit=20").ToArray();
-            Console.WriteLine($"Fetched {allApps.Length} apps.");
+            var allApps = restClient.FetchAllItems("app", 20);
             foreach (var app in allApps)
             {
+                appCnt++;
                 Console.WriteLine($"  - {app["id"]} : \"{app["name"]}\"");
             }
-        }
-
-        public static IEnumerable<JObject> FetchAll(IRestClientGeneric restClient, string endpoint)
-        {
-            var data = new List<JObject>();
-            var next = endpoint;
-            // Continue fetching data as long as there is a "next" reference.
-            while (next != null)
-            {
-                var spacesData = restClient.Get<JObject>(next);
-                data.AddRange(spacesData["data"].Values<JObject>());
-                next = spacesData["links"]["next"]?["href"]?.Value<string>()?.Substring(restClient.Url.Length);
-            }
-
-            return data;
-        }
+			Console.WriteLine($"Fetched {appCnt} apps.");
+		}
     }
 }