+++ Qlik.Sense.RestClient/Qlik.Sense.RestClient/RestClientQcs.cs
@@ -1,7 +1,9 @@
 ﻿using System.Security.Authentication;
 using System.Threading.Tasks;
 using System;
+using System.Collections.Generic;
 using System.Net;
+using Newtonsoft.Json.Linq;
 
 namespace Qlik.Sense.RestClient
 {
@@ -14,9 +16,45 @@ namespace Qlik.Sense.RestClient
         void AsExistingSessionViaQcs(QcsSessionInfo sessionInfo);
 
         IRestClientQcs WithContentType(string contentType);
-    }
 
-    public class RestClientQcs : RestClientGeneric, IRestClientQcs
+        /// <summary>
+        /// Utility method to fetch resources using the standard paging mechanism used in QCS. The endpoint is called
+        /// repeatedly as long as there is a <c>next</c> property in the response.
+        /// </summary>
+        /// <param name="endpoint">The initial endpoint used for page iteration.</param>
+        /// <returns>An <c>IEnumerable</c> with the set of resources fetched.</returns>
+        IEnumerable<JObject> FetchAll(string endpoint);
+
+		/// <summary>
+		/// Utility method to fetch all resources of a specific type. Method uses the <c>items</c> endpoint and calls
+		/// that endpoint repeatedly as long as there is a <c>next</c> property in the response.
+		/// </summary>
+		/// <example>var apps = FetchAllItems("app")</example>
+		/// <param name="resourceType">The type of the resource to fetch.</param>
+		/// <param name="pageSize">Number of items to fetch per REST call.  Must be in range 1-100.</param>
+		/// <returns>An <c>IEnumerable</c> with the set of resources fetched.</returns>
+		IEnumerable<JObject> FetchAllItems(string resourceType, int pageSize = 10);
+
+		/// <summary>
+		/// Utility method to fetch resources using the standard paging mechanism used in QCS. The endpoint is called
+		/// repeatedly as long as there is a <c>next</c> property in the response.
+		/// </summary>
+		/// <param name="endpoint">The initial endpoint used for page iteration.</param>
+		/// <returns>An <c>IEnumerable</c> with the set of resources fetched.</returns>
+        Task<IEnumerable<JObject>> FetchAllAsync(string endpoint);
+
+		/// <summary>
+		/// Utility method to fetch all resources of a specific type. Method uses the <c>items</c> endpoint and calls
+		/// that endpoint repeatedly as long as there is a <c>next</c> property in the response.
+		/// </summary>
+		/// <example>var apps = FetchAllItems("app")</example>
+		/// <param name="resourceType">The type of the resource to fetch.</param>
+		/// <param name="pageSize">Number of items to fetch per REST call. Must be in range 1-100.</param>
+		/// <returns>An <c>IEnumerable</c> with the set of resources fetched.</returns>
+		Task<IEnumerable<JObject>> FetchAllItemsAsync(string resourceType, int pageSize = 10);
+	}
+
+	public class RestClientQcs : RestClientGeneric, IRestClientQcs
     {
         private RestClientQcs(RestClientQcs source) : base(source)
         {
@@ -64,7 +102,57 @@ namespace Qlik.Sense.RestClient
             return client;
         }
 
-        private void AddBearerToken(string token)
+		/// <inheritdoc />
+		public IEnumerable<JObject> FetchAllItems(string resourceType, int pageSize = 10)
+		{
+			if (pageSize < 1 || pageSize > 100)
+				throw new ArgumentOutOfRangeException(nameof(pageSize), "Page size must be in the range 1-100.");
+
+	        return FetchAll($"/api/v1/items?resourceType={resourceType}&limit={pageSize}");
+        }
+
+        /// <inheritdoc />
+		public Task<IEnumerable<JObject>> FetchAllItemsAsync(string resourceType, int pageSize = 10)
+		{
+			if (pageSize < 1 || pageSize > 100)
+				throw new ArgumentOutOfRangeException(nameof(pageSize), "Page size must be in the range 1-100.");
+
+			return FetchAllAsync($"/api/v1/items?resourceType={resourceType}&limit={pageSize}");
+		}
+
+		/// <inheritdoc />
+		public IEnumerable<JObject> FetchAll(string endpoint)
+        {
+	        var next = endpoint;
+	        // Continue fetching data as long as there is a "next" reference.
+	        while (next != null)
+	        {
+		        var spacesData = Get<JObject>(next);
+		        foreach (var obj in spacesData["data"].Values<JObject>())
+		        {
+			        yield return obj;
+		        }
+		        next = spacesData["links"]["next"]?["href"]?.Value<string>()?.Substring(Url.Length);
+	        }
+        }
+
+		/// <inheritdoc />
+		public async Task<IEnumerable<JObject>> FetchAllAsync(string endpoint)
+		{
+			var data = new List<JObject>();
+			var next = endpoint;
+			// Continue fetching data as long as there is a "next" reference.
+			while (next != null)
+			{
+				var spacesData = await GetAsync<JObject>(next).ConfigureAwait(false);
+				data.AddRange(spacesData["data"].Values<JObject>());
+				next = spacesData["links"]["next"]?["href"]?.Value<string>()?.Substring(Url.Length);
+			}
+
+			return data;
+		}
+        
+		private void AddBearerToken(string token)
         {
             CustomHeaders.Add("Authorization", "Bearer " + token);
         }